requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.1
cycler==0.12.1
databricks-sdk==0.53.0
Deprecated==1.2.18
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
gto==1.7.2
h11==0.16.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
multidict==6.4.3
narwhals==1.39.1
networkx==3.4.2
numpy==2.2.5
omegaconf==2.3.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.3.2
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==80.7.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
torch --index-url https://download.pytorch.org/whl/cpu
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
wandb==0.19.11
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zc.lockfile==3.0.post1
zipp==3.21.0
GitPython
